+++ modules/snooker/src/test/ts/browser/HeaderOperationsTest.ts
@@ -205,6 +205,22 @@ UnitTest.test('HeaderOperationsTest', () => {
     );
 
     Assertions.checkOld(
+      'TINY-6666: Convert body to header (section)',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table>' +
+      '<thead><tr><td>A1</td><td>B1</td><td>C1</td><td>D1</td></tr></thead>' +
+      '<tbody><tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr></tbody>' +
+      '</table>',
+
+      '<table><tbody>' +
+      '<tr><td>A1</td><td>B1</td><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeRowHeader, 0, 0, 0, TableSection.section()
+    );
+
+    Assertions.checkOld(
       'TINY-6666: Convert body to footer (section)',
       Optional.some({ section: 1, row: 0, column: 1 }),
       '<table>' +
@@ -234,6 +250,22 @@ UnitTest.test('HeaderOperationsTest', () => {
       '</table>',
 
       TableOperations.makeRowFooter, 0, 0, 1, TableSection.sectionCells()
+    );
+
+    Assertions.checkOld(
+      'TINY-7709: Convert single row with colspan to header',
+      Optional.some({ section: 0, row: 0, column: 1 }),
+      '<table><tbody>' +
+      '<tr><th colspan="2" scope="colgroup">A1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><td colspan="2">A1</td><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeRowHeader, 0, 0, 1
     );
   };
 
@@ -534,7 +566,7 @@ UnitTest.test('HeaderOperationsTest', () => {
             '<td>D1</td>' +
           '</tr>' +
           '<tr>' +
-            '<td rowspan="2">A2</td>' +
+            '<td>A2</td>' +
             '<td>B2</td>' +
             '<td>C2</td>' +
             '<td>D2</td>' +
@@ -810,6 +842,24 @@ UnitTest.test('HeaderOperationsTest', () => {
 
       TableOperations.unmakeColumnHeader, 0, 0, 0
     );
+
+    Assertions.checkOld(
+      'TINY-7709: Convert single column with rowspan to header',
+      Optional.some({ section: 0, row: 2, column: 0 }),
+      '<table><tbody>' +
+      '<tr><th rowspan="2" scope="rowgroup">A1</th><td>B1</td><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><td rowspan="2">A1</td><td>B1</td><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeColumnHeader, 0, 2, 0
+    );
   };
 
   const testMultipleColumnHeader = () => {
@@ -1072,7 +1122,7 @@ UnitTest.test('HeaderOperationsTest', () => {
     );
 
     Assertions.checkOldMultiple(
-      'TINY-6765: Check that locked columns in the selection are not coverted to header columns',
+      'TINY-6765: Check that locked columns in the selection are not converted to header columns',
       Optional.some({ section: 0, row: 1, column: 0 }),
 
       generateTestTable(
@@ -1194,7 +1244,7 @@ UnitTest.test('HeaderOperationsTest', () => {
       '<tr><td>A1</td><td>B1</td><td>C1</td><td>D1</td></tr>' +
       '</thead>' +
       '<tbody>' +
-      '<tr><th scope="col">A2</th><td scope="col">B2</td><th scope="col">C2</th><th scope="col">D2</th></tr>' +
+      '<tr><th scope="col">A2</th><td>B2</td><th scope="col">C2</th><th scope="col">D2</th></tr>' +
       '</tbody></table>',
 
       '<table><thead>' +
@@ -1259,7 +1309,7 @@ UnitTest.test('HeaderOperationsTest', () => {
       '<tr><td>A1</td><td>B1</td><td>C1</td><td>D1</td></tr>' +
       '</thead>' +
       '<tbody>' +
-      '<tr><td scope="col">A2</td><td scope="col">B2</td><th scope="col">C2</th><th scope="col">D2</th></tr>' +
+      '<tr><td>A2</td><td>B2</td><th scope="col">C2</th><th scope="col">D2</th></tr>' +
       '</tbody></table>',
 
       '<table><thead>' +
@@ -1276,10 +1326,150 @@ UnitTest.test('HeaderOperationsTest', () => {
     );
   };
 
+  const testMixedHeaders = () => {
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert regular column to header column with an existing header row',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><tbody>' +
+      '<tr><th scope="row">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><th scope="row">A2</th><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><th scope="col">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeColumnsHeader, [
+        { section: 0, row: 0, column: 0 },
+        { section: 0, row: 1, column: 0 },
+        { section: 0, row: 2, column: 0 }
+      ]
+    );
+
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert header column to regular column with an existing header row',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><tbody>' +
+      '<tr><th scope="col">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><th scope="row">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><th scope="row">A2</th><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.unmakeColumnsHeader, [
+        { section: 0, row: 0, column: 0 },
+        { section: 0, row: 1, column: 0 },
+        { section: 0, row: 2, column: 0 }
+      ]
+    );
+
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert header column to regular column with an existing thead section header row',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><thead>' +
+      '<tr><td>A1</td><td>B1</td><th>C1</th><th>D1</th></tr>' +
+      '</thead><tbody>' +
+      '<tr><td>A2</td><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><thead>' +
+      '<tr><th scope="row">A1</th><td>B1</td><th>C1</th><th>D1</th></tr>' +
+      '</thead><tbody>' +
+      '<tr><th scope="row">A2</th><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.unmakeColumnsHeader, [
+        { section: 0, row: 0, column: 0 },
+        { section: 1, row: 0, column: 0 },
+        { section: 1, row: 1, column: 0 }
+      ]
+    );
+
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert regular row to header row with an existing header column',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><tbody>' +
+      '<tr><th scope="col">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>A2</td><th scope="row">B2</th><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><th scope="row">B3</th><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><td>A1</th><th scope="row">B1</th><td>C1</td><td>D1</th></tr>' +
+      '<tr><td>A2</td><th scope="row">B2</th><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><th scope="row">B3</th><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeRowsHeader, [
+        { section: 0, row: 0, column: 0 },
+        { section: 0, row: 0, column: 1 },
+        { section: 0, row: 0, column: 2 },
+        { section: 0, row: 0, column: 3 }
+      ]
+    );
+
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert header row to regular row with an existing header column',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><tbody>' +
+      '<tr><td>A1</td><th scope="row">B1</th><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>A2</td><th scope="row">B2</th><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><th scope="row">B3</th><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><th scope="col">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>A2</td><th scope="row">B2</th><td>C2</td><td>D2</td></tr>' +
+      '<tr><td>A3</td><th scope="row">B3</th><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeRowsBody, [
+        { section: 0, row: 0, column: 0 },
+        { section: 0, row: 0, column: 1 },
+        { section: 0, row: 0, column: 2 },
+        { section: 0, row: 0, column: 3 }
+      ]
+    );
+
+    Assertions.checkOldMultiple(
+      'TINY-7709: Convert header row with rowspan to regular row with an existing header column',
+      Optional.some({ section: 0, row: 0, column: 0 }),
+      '<table><tbody>' +
+      '<tr><th rowspan="2" scope="rowgroup">A1</th><td>B1</td><td>C1</td><td>D1</td></tr>' +
+      '<tr><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      '<table><tbody>' +
+      '<tr><th rowspan="2" scope="col">A1</th><th scope="col">B1</th><th scope="col">C1</th><th scope="col">D1</th></tr>' +
+      '<tr><td>B2</td><td>C2</td><td>D2</td></tr>' +
+      '<tr><th scope="row">A3</th><td>B3</td><td>C3</td><td>D3</td></tr>' +
+      '</tbody></table>',
+
+      TableOperations.makeRowsBody, [
+        { section: 0, row: 0, column: 0 },
+        { section: 0, row: 0, column: 1 },
+        { section: 0, row: 0, column: 2 },
+        { section: 0, row: 0, column: 3 }
+      ]
+    );
+  };
+
   testSingleRowSection();
   testMultipleRowSection();
   testSingleColumnHeader();
   testMultipleColumnHeader();
   testSingleCellHeader();
   testMultipleCellHeader();
+  testMixedHeaders();
 });